requirements.txt
amqp==2.5.2
arrow==0.15.5
billiard==3.6.2.0
boto3==1.11.9
botocore==1.14.9
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
docutils==0.15.2
future==0.18.2
importlib-metadata==1.5.0
jmespath==0.9.4
kombu==4.6.7
pkg-resources==0.0.0
pycparser==2.19
pycurl==7.43.0.5
python-dateutil==2.8.1
python-telegram-bot==12.3.0
pytz==2019.3
rutimeparser==1.1.1
s3transfer==0.3.2
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
vine==1.3.0
zipp==2.1.0
